requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
pydantic-settings
langchain
langchain-core
langchain-community
langchain-openai
langgraph
chromadb
sentence-transformers
rank_bm25
pdfplumber
python-docx
markdown
aiofiles
langchain_openai
